change size parameters to use 64-bit longs

+++ HashZeroBytes.cpp
@@ -11,10 +11,10 @@ int main( int argc, char* argv[] ){
         std::cout << "Usage: HashZeroBytes <N>" << std::endl;
         return 1;
     }
-    int messageLength = atoi( argv[1] );
+    long long messageLength = atoll( argv[1] );
     HashFunction *sha3 = new SHA3( DIGEST_BYTES );
     unsigned char digest[DIGEST_BYTES];
-    for( int i = 0; i < messageLength; i++ ){
+    for( long long i = 0; i < messageLength; i++ ){
         sha3->hash( 0 );
     }
     sha3->digest( digest );
